+++ SandboxiePlus/SandMan/SbiePlusAPI.h
@@ -68,7 +68,15 @@ public:
 	CSandBoxPlus(const QString& BoxName, class CSbieAPI* pAPI);
 	virtual ~CSandBoxPlus();
 
-	virtual QString			GetDisplayName() const;
+	enum EDisplayNameContext
+	{
+		eDisplayNormal,
+		eDisplayCompact
+	};
+
+	virtual QString			GetDisplayName(EDisplayNameContext Context = eDisplayNormal) const;
+	static QString			FormatDisplayName(const QString& BoxName, const QString& BoxAlias, int DisplayMode, EDisplayNameContext Context = eDisplayNormal);
+	virtual QString			GetBoxToolTip() const;
 
 	SB_PROGRESS				CopyBox(const QString& DestDir);
 
@@ -253,6 +261,7 @@ protected:
 	bool					m_NoForce;
 	QRgb					m_BoxColor;
 	QString					m_BoxAlias;
+	int						m_BoxAliasDisplayMode;
 };
 
 ///////////////////////////////////////////////////////////////////////////////